add type modue
